Derive traverse (#67)
* derive traverse in the @deriveFixedPoint macro annotation

* drop kittens dependency

* create a @deriveTraverse annotation

The use case for this one is that we won't always want to use the
@deriveFixedPoint annotation in a recursive data type.  We will
sometimes prefer to implement the fixedpoitn adt ourselves but don't
bother to create a `Traverse` instance that can be mechanically derived.

* fully qualify paths in codegen
