+++ modules/macros/src/main/scala/qq/droste/macros/annotations.scala
@@ -4,9 +4,14 @@ import scala.language.experimental.macros
 import scala.annotation.StaticAnnotation
 import scala.annotation.compileTimeOnly
 
-import impl.deriveFixedPointMacro
+import impl.Macros
 
 @compileTimeOnly("enable macro paradise to expand macro annotations")
 class deriveFixedPoint extends StaticAnnotation {
-  def macroTransform(annottees: Any*): Any = macro deriveFixedPointMacro.impl
+  def macroTransform(annottees: Any*): Any = macro Macros.deriveFixedPoint
+}
+
+@compileTimeOnly("enable macro paradise to expand macro annotations")
+class deriveTraverse extends StaticAnnotation {
+  def macroTransform(annottees: Any*): Any = macro Macros.deriveTraverse
 }
